clarify schema testing module
